pip-requirements.txt
ipykernel
matplotlib
pyyaml
pandas
openpyxl
halo
baostock
ipywidgets
torch